requirements.txt
# Data Ingestion
yfinance==0.2.36
kafka-python==2.0.2
requests==2.31.0
newsapi-python==0.2.7

# Data Processing
pyspark==3.3.2
pandas==2.1.0
numpy==1.24.3
delta-spark==2.3.0

# Machine Learning
scikit-learn==1.3.0
xgboost==2.0.0
joblib==1.3.2

# Database
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.0
sqlalchemy==1.4.49
psycopg2-binary==2.9.7

# DBT
dbt-core==1.6.0
dbt-snowflake==1.6.0

# Dashboard
streamlit==1.28.0
plotly==5.17.0
streamlit-autorefresh==1.0.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.4.0